+++ wuvt/templates/trackman/log.js
@@ -1,5 +1,5 @@
-var djset_id = {{ djset.id }};
-var dj_id = {{ djset.dj.id }};
+var djset_id = {{ djset_id or "null" }};
+var dj_id = {{ dj_id }};
 // Dictionary mapping rotation_ids to strings
 var rotations = {{ rotations|tojson|safe }};
